the first commit for my basic tic tac toe game
I will update later with the version that has comments, I need to backup quickly because my laptop crashes. I had to recreate this repo. Initially it was declaring winners when three dots fit the checker's logic which was wrong, I made adjustments to that. I created a new repo for other reasons